+++ jcnet/backwardFunctions/BackwardFunction.java
@@ -2,6 +2,6 @@ package jcnet.backwardfunctions;
 
 public abstract class BackwardFunction {
 	
-	public abstract void backward(float error);
+	public abstract void backward(double error);
 
 }
